// src/types.ts
export interface Product {
  id: string;
  sku: string;
  name: string;
  slug: string;
  description: string;
  price: number;
  offer_price: number | null;
  stock: number;
  category_id: string;
  category?: string;
  brand_id: string;
  featured: boolean;
  active: boolean;
  technical_sheet_url: string | null;
  barcode_qr: string | null;
  rating_stars?: number;
  rating_count?: number;
  cost_price?: number | null;
  unit?: string | null;
  units?: string | null;
  margin_1?: number | null;
  margin_2?: number | null;
  margin_3?: number | null;
  selected_margin_type?: 1 | 2 | 3 | string | number | null;
  tax_id?: string | null;
  tax_rate?: number | null;
  expiration_date?: string | null;
  critical_stock?: number | null;
  location?: string | null;
  created_at?: string;
  updated_at?: string;
}